Django_for_web/requirements.txt
asgiref==3.2.3
attrs==17.4.0
Automat==0.6.0
beautifulsoup4==4.8.2
bleach==3.1.1
bs4==0.0.1
certifi==2017.11.5
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
cssselect==1.0.3
cycler==0.10.0

django-pagination-bootstrap==1.2.4
docutils==0.16
hyperlink==17.3.1
idna==2.6
importlib-metadata==1.5.0
incremental==17.5.0


kiwisolver==1.1.0
lxml==4.5.0
matplot==0.1.9

mysql==0.0.2
mysql-connector-python==8.0.19
mysqlclient==1.4.6
numpy==1.18.1
olefile==0.44

parsel==1.5.2
Pillow==7.0.0
pkginfo==1.5.0.1
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.6.1
PyHamcrest==2.0.1
pyloco==0.0.139
pymysql==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
pywin32-ctypes==0.2.0
queuelib==1.5.0
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
scipy==1.4.1
Scrapy==1.8.0
service-identity==18.1.0
SimpleWebSocketServer==0.1.1
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.1
tqdm==4.43.0

Twisted==19.10.0
typing==3.7.4.1
unicode==2.7
urllib3==1.25.8
ushlex==0.99.1
w3lib==1.21.0
webencodings==0.5.1
websocket-client==0.57.0
wordcloud==1.6.0
